Refactor Poseidon default traits and structs
