+++ algorithms/src/crypto_hash/mod.rs
@@ -17,9 +17,6 @@
 pub mod blake2xs;
 pub use blake2xs::*;
 
-pub mod grain_lfsr;
-pub use grain_lfsr::*;
-
 pub mod poseidon;
 pub use poseidon::*;
