+++ algorithms/src/crypto_hash/poseidon.rs
@@ -14,122 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{
-    crypto_hash::PoseidonGrainLFSR,
-    traits::{AlgebraicSponge, DefaultCapacityAlgebraicSponge, DuplexSpongeMode},
-    CryptoHash,
-};
-use snarkvm_fields::{Fp256, Fp256Parameters, Fp384, Fp384Parameters, PoseidonDefaultParameters, PrimeField};
-use snarkvm_utilities::{FromBytes, ToBytes};
+use crate::{AlgebraicSponge, CryptoHash, DefaultCapacityAlgebraicSponge, DuplexSpongeMode};
+use snarkvm_fields::{PoseidonDefaultField, PoseidonParameters, PrimeField};
 
 use smallvec::SmallVec;
 use std::{
-    io::{Read, Result as IoResult, Write},
     ops::{Index, IndexMut, Range},
     sync::Arc,
 };
-
-/// Parameters and RNG used
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct PoseidonParameters<F: PrimeField, const RATE: usize, const CAPACITY: usize> {
-    /// number of rounds in a full-round operation
-    pub full_rounds: usize,
-    /// number of rounds in a partial-round operation
-    pub partial_rounds: usize,
-    /// Exponent used in S-boxes
-    pub alpha: u64,
-    /// Additive Round keys. These are added before each MDS matrix application to make it an affine shift.
-    /// They are indexed by `ark[round_num][state_element_index]`
-    pub ark: Vec<Vec<F>>,
-    /// Maximally Distance Separating Matrix.
-    pub mds: Vec<Vec<F>>,
-}
-
-impl<F: PrimeField, const RATE: usize, const CAPACITY: usize> PoseidonParameters<F, RATE, CAPACITY> {
-    /// Initialize the parameter for Poseidon Sponge.
-    pub fn new(full_rounds: usize, partial_rounds: usize, alpha: u64, mds: Vec<Vec<F>>, ark: Vec<Vec<F>>) -> Self {
-        assert_eq!(ark.len(), full_rounds + partial_rounds);
-        for item in &ark {
-            assert_eq!(item.len(), RATE + CAPACITY);
-        }
-        assert_eq!(mds.len(), RATE + CAPACITY);
-        for item in &mds {
-            assert_eq!(item.len(), RATE + CAPACITY);
-        }
-        Self { full_rounds, partial_rounds, alpha, mds, ark }
-    }
-}
-
-impl<F: PrimeField, const RATE: usize, const CAPACITY: usize> ToBytes for PoseidonParameters<F, RATE, CAPACITY> {
-    #[inline]
-    fn write_le<W: Write>(&self, mut writer: W) -> IoResult<()> {
-        (self.full_rounds as u32).write_le(&mut writer)?;
-        (self.partial_rounds as u32).write_le(&mut writer)?;
-        self.alpha.write_le(&mut writer)?;
-
-        (self.ark.len() as u32).write_le(&mut writer)?;
-        for fields in &self.ark {
-            (fields.len() as u32).write_le(&mut writer)?;
-            for field in fields {
-                field.write_le(&mut writer)?;
-            }
-        }
-
-        (self.mds.len() as u32).write_le(&mut writer)?;
-        for fields in &self.mds {
-            (fields.len() as u32).write_le(&mut writer)?;
-            for field in fields {
-                field.write_le(&mut writer)?;
-            }
-        }
-
-        (RATE as u32).write_le(&mut writer)?;
-        (CAPACITY as u32).write_le(&mut writer)
-    }
-}
-
-impl<F: PrimeField, const RATE: usize, const CAPACITY: usize> FromBytes for PoseidonParameters<F, RATE, CAPACITY> {
-    #[inline]
-    fn read_le<R: Read>(mut reader: R) -> IoResult<Self> {
-        let full_rounds: u32 = FromBytes::read_le(&mut reader)?;
-        let partial_rounds: u32 = FromBytes::read_le(&mut reader)?;
-        let alpha: u64 = FromBytes::read_le(&mut reader)?;
-
-        let ark_length: u32 = FromBytes::read_le(&mut reader)?;
-        let mut ark = Vec::with_capacity(ark_length as usize);
-        for _ in 0..ark_length {
-            let num_fields: u32 = FromBytes::read_le(&mut reader)?;
-            let mut fields = Vec::with_capacity(num_fields as usize);
-
-            for _ in 0..num_fields {
-                let field: F = FromBytes::read_le(&mut reader)?;
-                fields.push(field);
-            }
-            ark.push(fields);
-        }
-
-        let mds_length: u32 = FromBytes::read_le(&mut reader)?;
-        let mut mds = Vec::with_capacity(mds_length as usize);
-        for _ in 0..mds_length {
-            let num_fields: u32 = FromBytes::read_le(&mut reader)?;
-            let mut fields = Vec::with_capacity(num_fields as usize);
-
-            for _ in 0..num_fields {
-                let field: F = FromBytes::read_le(&mut reader)?;
-                fields.push(field);
-            }
-            mds.push(fields);
-        }
-
-        let rate: u32 = FromBytes::read_le(&mut reader)?;
-        let capacity: u32 = FromBytes::read_le(&mut reader)?;
-        if rate != RATE as u32 || capacity != CAPACITY as u32 {
-            return Err(std::io::ErrorKind::Other.into());
-        }
-
-        Ok(Self::new(full_rounds as usize, partial_rounds as usize, alpha, mds, ark))
-    }
-}
 
 #[derive(Copy, Clone, Debug)]
 pub struct State<F: PrimeField, const RATE: usize, const CAPACITY: usize> {
@@ -445,113 +337,3 @@ impl<F: PrimeField + PoseidonDefaultField, const RATE: usize, const OPTIMIZED_FO
         &self.parameters
     }
 }
-
-impl<F: PrimeField + PoseidonDefaultField, const RATE: usize, const OPTIMIZED_FOR_WEIGHTS: bool>
-    From<PoseidonParameters<F, RATE, 1>> for PoseidonCryptoHash<F, RATE, OPTIMIZED_FOR_WEIGHTS>
-{
-    fn from(parameters: PoseidonParameters<F, RATE, 1>) -> Self {
-        Self { parameters: Arc::new(parameters) }
-    }
-}
-
-impl<F: PrimeField + PoseidonDefaultField, const RATE: usize, const OPTIMIZED_FOR_WEIGHTS: bool>
-    From<Arc<PoseidonParameters<F, RATE, 1>>> for PoseidonCryptoHash<F, RATE, OPTIMIZED_FOR_WEIGHTS>
-{
-    fn from(parameters: Arc<PoseidonParameters<F, RATE, 1>>) -> Self {
-        Self { parameters }
-    }
-}
-
-impl<F: PrimeField + PoseidonDefaultField, const RATE: usize, const OPTIMIZED_FOR_WEIGHTS: bool> ToBytes
-    for PoseidonCryptoHash<F, RATE, OPTIMIZED_FOR_WEIGHTS>
-{
-    #[inline]
-    fn write_le<W: Write>(&self, mut writer: W) -> IoResult<()> {
-        self.parameters.write_le(&mut writer)
-    }
-}
-
-impl<F: PrimeField + PoseidonDefaultField, const RATE: usize, const OPTIMIZED_FOR_WEIGHTS: bool> FromBytes
-    for PoseidonCryptoHash<F, RATE, OPTIMIZED_FOR_WEIGHTS>
-{
-    #[inline]
-    fn read_le<R: Read>(mut reader: R) -> IoResult<Self> {
-        let parameters: PoseidonParameters<F, RATE, 1> = FromBytes::read_le(&mut reader)?;
-        Ok(Self::from(parameters))
-    }
-}
-
-/// A field with Poseidon parameters associated
-pub trait PoseidonDefaultField: PrimeField {
-    /// Obtain the default Poseidon parameters for this rate and for this prime field,
-    /// with a specific optimization goal.
-    fn default_poseidon_parameters<const RATE: usize>(
-        optimized_for_weights: bool,
-    ) -> Option<PoseidonParameters<Self, RATE, 1>> {
-        /// Internal function that computes the ark and mds from the Poseidon Grain LFSR.
-        fn find_poseidon_ark_and_mds<F: PrimeField, const RATE: usize>(
-            full_rounds: u64,
-            partial_rounds: u64,
-            skip_matrices: u64,
-        ) -> (Vec<Vec<F>>, Vec<Vec<F>>) {
-            let mut lfsr =
-                PoseidonGrainLFSR::new(false, F::size_in_bits() as u64, (RATE + 1) as u64, full_rounds, partial_rounds);
-
-            let mut ark = Vec::<Vec<F>>::new();
-            for _ in 0..(full_rounds + partial_rounds) {
-                ark.push(lfsr.get_field_elements_rejection_sampling(RATE + 1));
-            }
-
-            let mut mds = vec![vec![F::zero(); RATE + 1]; RATE + 1];
-            for _ in 0..skip_matrices {
-                let _ = lfsr.get_field_elements_mod_p::<F>(2 * (RATE + 1));
-            }
-
-            // a qualifying matrix must satisfy the following requirements
-            // - there is no duplication among the elements in x or y
-            // - there is no i and j such that x[i] + y[j] = p
-            // - the resultant MDS passes all the three tests
-
-            let xs = lfsr.get_field_elements_mod_p::<F>(RATE + 1);
-            let ys = lfsr.get_field_elements_mod_p::<F>(RATE + 1);
-
-            for (i, x) in xs.iter().enumerate().take(RATE + 1) {
-                for (j, y) in ys.iter().enumerate().take(RATE + 1) {
-                    mds[i][j] = (*x + y).inverse().unwrap();
-                }
-            }
-
-            (ark, mds)
-        }
-
-        let default_entries = match optimized_for_weights {
-            true => Self::Parameters::PARAMS_OPT_FOR_WEIGHTS,
-            false => Self::Parameters::PARAMS_OPT_FOR_CONSTRAINTS,
-        };
-
-        default_entries.iter().find(|entry| entry.rate == RATE).map(|entry| {
-            let (ark, mds) = find_poseidon_ark_and_mds::<Self, RATE>(
-                entry.full_rounds as u64,
-                entry.partial_rounds as u64,
-                entry.skip_matrices as u64,
-            );
-
-            PoseidonParameters {
-                full_rounds: entry.full_rounds,
-                partial_rounds: entry.partial_rounds,
-                alpha: entry.alpha as u64,
-                ark,
-                mds,
-            }
-        })
-    }
-}
-
-macro_rules! impl_poseidon_default_parameters_field {
-    ($field: ident, $params: ident) => {
-        impl<P: $params + PoseidonDefaultParameters> PoseidonDefaultField for $field<P> {}
-    };
-}
-
-impl_poseidon_default_parameters_field!(Fp256, Fp256Parameters);
-impl_poseidon_default_parameters_field!(Fp384, Fp384Parameters);
